COPYRIGHT.md
This is the work of 'Petros Soutzis' and is protected by the 'Copyright Law' (©).

This was a course assignment for Lancaster University, Computer Science undergrad. degree.
You may download, read and run the program for educational purposes only.
Using this code for anything but educational purposes (e.g. commercial use), without my knowledge is strictly prohibited.
Parsing this work as your own, or plagiarizing and embedding this to your own code is prohibited.

-Contact me at 'skata456@hotmail.com'